PKCS#11 对象句柄编码限制在 32 位内；断言宏改用 long long 比较
两处都与 32 位 ABI 有关，在 64 位平台上行为不变。

一、PKCS#11 对象句柄编码

CK_OBJECT_HANDLE 就是 CK_ULONG，而 CK_ULONG 的定义是 unsigned long：
x86_64 与 aarch64 上是 8 字节，32 位 ABI（如 armv7l）上只有 4 字节。
原编码用 1ULL << 63 作公钥位、1ULL << 62 作会话密钥位、generation << 32
带代次，这三者在 32 位上全部被赋值截断为 0，后果是：

  · C_SignInit 会接受公钥句柄去签名 —— 判据是 hKey & PUB_BIT，而那一位没了
  · KEM 会话密钥对象无法寻址
  · generation 从句柄里消失，密钥重生成后的旧句柄失效检查形同虚设

改为整体落在 32 位内的编码：bit 31 公钥、bit 30 会话密钥、
bits 12-29 槽位 generation 的低 18 位、bits 0-11 索引。
与核心层 64 位 hsm_handle_t 之间的转换收进 p11_handle_of_core() 与
core_handle_of() 两个函数，原先散落各处的 (hsm_handle_t)(h & ~PUB_BIT)
一并换掉；core_handle_of() 顺带做索引与 generation 的校验，
过期句柄在进核心层之前就被拒。

generation 只带 18 位的代价写在注释里：同一槽位重生成 2^18 次之后旧句柄的
这 18 位会重新对上，核心层仍做完整的 64 位比较，截断只影响本模块的前置检查。

tests/unit/test_p11.c 里原本硬写了一份 1ULL << 63，同步改为与模块一致的常量。

二、tests/testlib.h 的 CHECK_EQ_INT

原先经 long 比较，在 32 位 ABI 上会把 64 位实参截断，于是"高 32 位不同、
低 32 位相同"的两个值被判为相等，断言变成假通过。改为 long long。
当前没有调用点受此影响，属于潜在问题。

ctest 45/45，ASan+UBSan 45/45，test_p11 254 条断言全部通过。

+++ tests/testlib.h
@@ -24,12 +24,15 @@ static const char *g_case = "";
 	}                                                                   \
 } while (0)
 
+/* 比较用 long long 而不是 long：long 在 32 位 ABI（armv7l）上只有 4 字节，
+ * 拿它去接 uint64_t 的句柄、计数器、时间戳会被截断，于是"高 32 位不同、
+ * 低 32 位相同"的两个值会被判成相等 —— 断言变成假通过。 */
 #define CHECK_EQ_INT(a, b) do {                                             \
-	long _a = (long)(a), _b = (long)(b);                                \
+	long long _a = (long long)(a), _b = (long long)(b);                 \
 	g_checks++;                                                         \
 	if (_a != _b) {                                                     \
 		g_fails++;                                                  \
-		fprintf(stderr, "FAIL %s:%d [%s] %s == %s (%ld vs %ld)\n",  \
+		fprintf(stderr, "FAIL %s:%d [%s] %s == %s (%lld vs %lld)\n", \
 		        __FILE__, __LINE__, g_case, #a, #b, _a, _b);        \
 	}                                                                   \
 } while (0)
